+++ Makefile
@@ -6,50 +6,64 @@ CFLAGS=-g -Wall -Wextra -Werror
 LFLAGS=-I includes/
 
 SRC_FILES =		main.c \
+				shell/shell.c \
+				shell/shell_utils.c \
+				shell/shell_utils_2.c \
+				shell/shell_utils_3.c \
+				builtin/ft_env.c \
+				builtin/ft_cd/ft_cd.c \
+				builtin/ft_cd/ft_cd_utils.c \
+				builtin/ft_echo.c \
+				builtin/ft_pwd.c \
+				builtin/ft_unset.c \
+				builtin/ft_export/ft_export.c \
+				builtin/ft_export/ft_export_utils.c \
+				builtin/ft_export/ft_export_utils_2.c \
+				builtin/ft_exit.c \
+				parsing/backslash.c \
+				parsing/braces.c \
+				parsing/embrace_expansion.c \
+				parsing/expansion.c \
+				parsing/parser.c \
+				parsing/quotes.c \
+				parsing/quotes_utils.c \
+				parsing/split_non_escaped.c \
+				validator/validator.c \
+				validator/validator_utils.c \
 				get_next_line.c \
-				parser.c \
 				search.c \
-				ft_env.c \
-				ft_cd.c \
-				ft_echo.c \
-				ft_pwd.c \
-				ft_unset.c \
-				ft_export.c \
-				ft_exit.c \
 				signals.c \
 				errors.c \
 				clean.c \
-				expansion.c \
-				embrace_expansion.c \
 				utils.c \
-				quotes.c \
-				backslash.c \
-				quotes_utils.c \
-				split_non_escaped.c \
 				redirections.c \
 				pipes.c \
 				permissions.c \
-				ft_shlvl.c \
-				braces.c
+				shlvl.c \
 			
 OBJ_FILES=$(SRC_FILES:.c=.o)
-SRC_DIR=srcs
-OBJ_DIR=objs
+SRC_DIR=srcs/
+OBJ_DIR=objs/
 LIB_DIR=libft
-SRC=$(SRC_FILES:%=$(SRC_DIR)/%)
-OBJS=$(OBJ_FILES:%=$(OBJ_DIR)/%)
+OBJS=$(addprefix $(OBJ_DIR), $(OBJ_FILES))
 
 
 all: 			$(OBJ_DIR) $(NAME)
 
 $(OBJ_DIR):		
 				@mkdir -p $(OBJ_DIR)
+				@mkdir -p $(OBJ_DIR)/shell
+				@mkdir -p $(OBJ_DIR)/builtin
+				@mkdir -p $(OBJ_DIR)/builtin/ft_cd
+				@mkdir -p $(OBJ_DIR)/builtin/ft_export
+				@mkdir -p $(OBJ_DIR)/parsing
+				@mkdir -p $(OBJ_DIR)/validator
 
 $(NAME): 		$(OBJS)
 				@make -C $(LIB_DIR)
 				@$(CC) $(CFLAGS) $(LFLAGS) $(OBJS) -Llibft -lft -o $(NAME)
 
-$(OBJ_DIR)/%.o:	$(SRC_DIR)/%.c
+$(OBJ_DIR)%.o:	$(SRC_DIR)%.c
 				@$(CC) $(LFLAGS) $(CFLAGS) -c $< -o $@
 
 clean:
