+++ includes/minishell.h
@@ -90,12 +90,11 @@ char	*set_opening_braces(char *str);
 char	*set_closing_braces(char *str);
 
 /*
-** Builtin functions and utils
+** Builtin functions
 */
 
 int		ft_env(t_shell *shell, char **args);
 int		ft_cd(t_shell *shell);
-int		is_symbolic_link(char *path);
 int		ft_echo(char **args);
 int		ft_pwd(t_shell *shell);
 int		ft_unset(t_shell *shell);
@@ -103,10 +102,71 @@ void	ft_exit(t_shell *shell);
 int 	ft_export(t_shell *shell, char *last_arg);
 
 /*
+** Builtin-cd utils functions
+*/
+
+void	set_path(char *path, const char *key, t_shell *shell);
+void	dir_err(t_shell *shell, char *path, char *err);
+void	change_dir(char *path, t_shell *shell);
+void	change_dir_error(t_shell *shell, char *path);
+short	home_not_exists(t_shell *shell, char *path);
+void	update_olddir(t_shell *shell, char *oldpwd);
+short	many_arguments(t_shell *shell);
+int		is_symbolic_link(char *path);
+
+/*
+** Builtin-export utils functions
+*/
+
+size_t  is_bigger(char *s1, char *s2);
+char	**sort_alpha(char **envp);
+short	is_escapable_char_export(char c);
+int		count_special_chars(char *str);
+char	*add_quotes_and_backslash(char *str);
+int     ft_strncmp_equal(char *str1, char *str2, int len);
+short	is_num(char c);
+short	is_forbidden_char(char *str, int i);
+short	has_valid_name(char *str);
+short	not_valid_keyname(t_shell *shell, char *var, char *command);
+char	**process_envs(char **env);
+size_t	print_sorted_env(t_shell *shell);
+void	export_values(t_shell *shell, char *last_arg, int j);
+
+/*
 ** Signal functions
 */
 
 void	signal_handler_waiting(int signal);
 void	signal_handler_running(int signal);
+
+/*
+** Shell functions
+*/
+
+void	minishell(char *line, t_shell *shell);
+char	*last_arg(char *arg);
+char	*update_last_arg(char **args);
+char	*append_pwd(char *value);
+char	*to_lower(char *input);
+int		exec_export(t_shell *shell);
+int		exec_env(t_shell *shell);
+char	*get_var_key(char *line, int it);
+short	contain_spaces(char *value);
+char	*add_space(char *line, char *output, size_t *it, char redir);
+size_t	additional_spaces(char *line);
+char	*inject_spaces(char *line);
+char	*safe_allocate(char *line, char *(*f)(char *));
+
+/*
+** Validator functions
+*/
+
+int		validate_input(t_shell *shell, char *line);
+char	*post_to_token(t_shell *shell, char *line, int it, char token);
+short	prior_to_token(char *line, int it, char token);
+short	nothing_after_pipe(char *line, int it);
+int		get_validator_err(t_shell *shell,
+							char *line, char separator, int it);
+void	safe_print_err(t_shell *shell, char *key);
 
 #endif
