+++ srcs/builtin/ft_unset.c
@@ -12,63 +12,20 @@
 
 #include "minishell.h"
 
-static short	is_num(char c)
-{
-	if (c >= '0' && c <= '9')
-		return (1);
-	return (0);
-}
 
-static short	is_forbidden_char(char *str, int i)
-{
-	if (str[i] == ' ' ||
-		str[i] == '\'' ||
-		str[i] == '\"' ||
-		str[i] == '$' ||
-		str[i] == '!' ||
-		str[i] == '|' ||
-		str[i] == ';' ||
-		str[i] == '&' ||
-		str[i] == '@')
-		return (1);
-	else if (str[i] == '\\' &&
-			(str[i + 1] != '_' && !is_num(str[i + 1])))
-		return (1);
-	return (0);
-}
-
-static short	has_valid_name(char *str)
+static short	has_valid_name_unset(char *str)
 {
 	int	i;
 
 	i = 0;
-	if (!*str)
+	if (!*str || is_num(*str) || *str == '=')
 		return (0);
-	while (str[i] && str[i] != '=')
+	while (str[i])
 	{
-		if (is_forbidden_char(str, i))
+		if (is_forbidden_char(str, i) || str[i] == '=')
 			return (0);
 		i++;
 	}
-	if (str[i] == '=' && !str[i + 1])
-		return (0);
-	return (1);
-}
-
-static	short	not_valid_keyname(t_shell *shell, char *var)
-{
-	char	*err_mssg;
-	int		len;
-
-	len = ft_strlen(var);
-	if (!(err_mssg = malloc((len + 28) * sizeof(char))))
-		return (0);
-	ft_strlcpy(err_mssg, " `", 3);
-	ft_strlcpy(err_mssg + 2, var, len + 1);
-	ft_strlcpy(err_mssg + 2 + len, "': not a valid identifier", 26);
-	shell->stat_loc = 1;
-	print_errors(shell, err_mssg, "unset");
-	free(err_mssg);
 	return (1);
 }
 
@@ -79,10 +36,10 @@ int				ft_unset(t_shell *shell)
 	it = 1;
 	while (shell->args[it])
 	{
-		if (has_valid_name(shell->args[it]))
+		if (has_valid_name_unset(shell->args[it]))
 			delete_environment(shell, shell->args[it], shell->env);
 		else
-			not_valid_keyname(shell, shell->args[it]);
+			not_valid_keyname(shell, shell->args[it], "unset");
 		it++;
 	}
 	return (1);
